backend/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
newspaper3k
pydantic
sqlite3  # built‑in, just for clarity